moved isDev to common
