+++ src/main/utils/common.ts
@@ -1,0 +1,2 @@
+import { app } from 'electron';
+export const isDev = !app.isPackaged;
